Refactor controller file to routes directory with one file per route type
